MLIBZ-2273 Refactor SqliteKeyValueStorePersister into a generic SqlKeyValuePersister, which works a specific SQL module. Refactor WebSqlKeyValueStorePersister to be an SQL module. Now WebSQL and SQLite both use the same persister, with different SQL modules.
