+++ src/core/datastore/persisters/browser/index.js
@@ -1,3 +1,3 @@
-export * from './websql-key-value-store-persister';
+export * from './websql-sql-module';
 export * from './indexeddb-key-value-store-persister';
 export * from './browser-key-value-persister';
