+++ src/core/datastore/persisters/index.js
@@ -3,4 +3,4 @@ export * from './key-value-persister';
 export * from './memory-key-value-persister';
 export * from './key-value-store-persister';
 export * from './browser';
-export * from './sqlite-key-value-store-persister';
+export * from './sql-key-value-store-persister';
