+++ src/html5/offline-data-storage.js
@@ -1,15 +1,19 @@
+import { Html5Client } from './client';
+
 import {
   repositoryProvider,
   storageType,
   KeyValueStoreOfflineRepository,
+  SqlKeyValueStorePersister,
   InmemoryOfflineRepository,
-  WebSqlKeyValueStorePersister,
+  WebSqlSqlModule,
   IndexedDbKeyValueStorePersister,
   BrowserKeyValuePersister
 } from '../core/datastore';
 
 const webSqlBuilder = (queue) => {
-  const persister = new WebSqlKeyValueStorePersister();
+  const sqlModule = new WebSqlSqlModule(Html5Client.sharedInstance().appKey);
+  const persister = new SqlKeyValueStorePersister(sqlModule);
   return new KeyValueStoreOfflineRepository(persister, queue);
 };
 
@@ -28,7 +32,6 @@ const sessionStorageBuilder = (queue) => {
   return new InmemoryOfflineRepository(persister, queue);
 };
 
-// TODO: this will grow, refactor
 const repoConstructors = {
   [storageType.webSql]: webSqlBuilder,
   [storageType.indexedDb]: indexedDbBuilder,
