+++ src/nativescript/offline-data-storage.ts
@@ -2,14 +2,14 @@ import {
   repositoryProvider,
   storageType,
   KeyValueStoreOfflineRepository,
-  SqliteKeyValueStorePersister
+  SqlKeyValueStorePersister
 } from '../core/datastore';
 import { Client } from './client';
-import { NativescriptSqliteModule } from './nativescript-sqlite-module';
+import { NativescriptSqlModule } from './nativescript-sql-module';
 
 const sqliteStorageBuilder = (queue) => {
-  const sqlModule = new NativescriptSqliteModule(Client.sharedInstance().appKey);
-  const persister = new SqliteKeyValueStorePersister(sqlModule);
+  const sqlModule = new NativescriptSqlModule(Client.sharedInstance().appKey);
+  const persister = new SqlKeyValueStorePersister(sqlModule);
   return new KeyValueStoreOfflineRepository(persister, queue);
 };
